Query: Combine QueryingEnumerable for regular and non composed FromSql case
